pyproject.toml
[project]
name = "phidata"
version = "2.4.10"
description = "Memory, knowledge and tools for LLMs."
requires-python = ">=3.7"
readme = "README.md"
authors = [
  {name = "Ashpreet Bedi", email = "ashpreet@phidata.com"}
]

dependencies = [
  "gitpython",
  "httpx",
  "pydantic",
  "pydantic-settings",
  "python-dotenv",
  "pyyaml",
  "rich",
  "tomli",
  "typer",
  "typing-extensions",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pytest",
    "ruff",
    "types-pyyaml"
]

docker = [
    "docker"
]

aws = [
    "docker",
    "boto3"
]

k8s = [
    "docker",
    "kubernetes"
]

[project.scripts]
phi = "phi.cli.entrypoint:phi_cli"

[project.urls]
homepage = "https://phidata.com"
documentation = "https://docs.phidata.com"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["phi*"]

[tool.setuptools.package-data]
phi = ["py.typed"]

[tool.pytest.ini_options]
testpaths = "tests"

[tool.ruff]
line-length = 120
exclude = ["phienv*", "aienv*"]
# Ignore `F401` (import violations) in all `__init__.py` files
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"phi/k8s/app/traefik/crds.py" = ["E501"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = ["phienv*", "aienv*", "scratch*", "wip*", "tmp*", "cookbook/examples/*"]

[[tool.mypy.overrides]]
module = [
  "altair.*",
  "arxiv.*",
  "anthropic.*",
  "apify_client.*",
  "boto3.*",
  "botocore.*",
  "bs4.*",
  "cohere.*",
  "docker.*",
  "duckdb.*",
  "exa_py.*",
  "firecrawl.*",
  "duckduckgo_search.*",
  "groq.*",
  "kubernetes.*",
  "lancedb.*",
  "langchain.*",
  "langchain_core.*",
  "llama_index.*",
  "mistralai.*",
  "newspaper.*",
  "nest_asyncio.*",
  "numpy.*",
  "ollama.*",
  "openai.*",
  "openbb.*",
  "pandas.*",
  "pinecone.*",
  "pyarrow.*",
  "pgvector.*",
  "psycopg.*",
  "pypdf.*",
  "qdrant_client.*",
  "rapidocr_onnxruntime.*",
  "requests.*",
  "simplejson.*",
  "serpapi.*",
  "setuptools.*",
  "sqlalchemy.*",
  "streamlit.*",
  "tavily.*",
  "textract.*",
  "vertexai.*",
  "voyageai.*",
  "wikipedia.*",
  "yfinance.*",
  "youtube_transcript_api.*",
]

ignore_missing_imports = true
